cleanup and resume_foreground
