+++ src/tool/cpp/cppwinrt/strings/base_coroutine.h
@@ -67,239 +67,6 @@ namespace winrt::Windows::Foundation
 
 namespace winrt
 {
-    [[nodiscard]] inline auto resume_background()
-    {
-        struct awaitable
-        {
-            bool await_ready() const noexcept
-            {
-                return false;
-            }
-
-            void await_resume() const noexcept
-            {
-            }
-
-            void await_suspend(std::experimental::coroutine_handle<> handle) const
-            {
-                if (!WINRT_TrySubmitThreadpoolCallback(callback, handle.address(), nullptr))
-                {
-                    throw_last_error();
-                }
-            }
-
-        private:
-
-            static void WINRT_CALL callback(void*, void* context) noexcept
-            {
-                std::experimental::coroutine_handle<>::from_address(context)();
-            }
-        };
-
-        return awaitable{};
-    }
-
-    template <typename T>
-    [[nodiscard]] auto resume_background(T&& context)
-    {
-        struct awaitable
-        {
-            awaitable(T&& context) : m_context(std::forward<T>(context))
-            {
-            }
-
-            bool await_ready() const noexcept
-            {
-                return false;
-            }
-
-            void await_resume() const noexcept
-            {
-            }
-
-            void await_suspend(std::experimental::coroutine_handle<> resume)
-            {
-                m_resume = resume;
-
-                if (!WINRT_TrySubmitThreadpoolCallback(callback, this, nullptr))
-                {
-                    throw_last_error();
-                }
-            }
-
-        private:
-
-            static void WINRT_CALL callback(void*, void* context) noexcept
-            {
-                auto that = static_cast<awaitable*>(context);
-                auto guard = that->m_context();
-                that->m_resume();
-            }
-
-            T&& m_context;
-            std::experimental::coroutine_handle<> m_resume{ nullptr };
-        };
-
-        return awaitable{ std::forward<T>(context) };
-    }
-
-    struct apartment_context
-    {
-        apartment_context()
-        {
-            m_context.capture(WINRT_CoGetObjectContext);
-        }
-
-        bool await_ready() const noexcept
-        {
-            return false;
-        }
-
-        void await_resume() const noexcept
-        {
-        }
-
-        void await_suspend(std::experimental::coroutine_handle<> handle) const
-        {
-            impl::com_callback_args args{};
-            args.data = handle.address();
-            check_hresult(m_context->ContextCallback(callback, &args, guid_of<impl::ICallbackWithNoReentrancyToApplicationSTA>(), 5, nullptr));
-        }
-
-    private:
-
-        static int32_t WINRT_CALL callback(impl::com_callback_args* args) noexcept
-        {
-            std::experimental::coroutine_handle<>::from_address(args->data)();
-            return impl::error_ok;
-        }
-
-        com_ptr<impl::IContextCallback> m_context;
-    };
-
-    [[nodiscard]] inline auto resume_after(Windows::Foundation::TimeSpan duration)
-    {
-        struct awaitable
-        {
-            explicit awaitable(Windows::Foundation::TimeSpan duration) noexcept :
-                m_duration(duration)
-            {
-            }
-
-            bool await_ready() const noexcept
-            {
-                return m_duration.count() <= 0;
-            }
-
-            void await_suspend(std::experimental::coroutine_handle<> handle)
-            {
-                m_timer.attach(check_pointer(WINRT_CreateThreadpoolTimer(callback, handle.address(), nullptr)));
-                int64_t relative_count = -m_duration.count();
-                WINRT_SetThreadpoolTimer(m_timer.get(), &relative_count, 0, 0);
-            }
-
-            void await_resume() const noexcept
-            {
-            }
-
-        private:
-
-            static void WINRT_CALL callback(void*, void* context, void*) noexcept
-            {
-                std::experimental::coroutine_handle<>::from_address(context)();
-            }
-
-            struct timer_traits
-            {
-                using type = impl::ptp_timer;
-
-                static void close(type value) noexcept
-                {
-                    WINRT_CloseThreadpoolTimer(value);
-                }
-
-                static constexpr type invalid() noexcept
-                {
-                    return nullptr;
-                }
-            };
-
-            handle_type<timer_traits> m_timer;
-            Windows::Foundation::TimeSpan m_duration;
-        };
-
-        return awaitable{ duration };
-    }
-
-    [[nodiscard]] inline auto resume_on_signal(void* handle, Windows::Foundation::TimeSpan timeout = {})
-    {
-        struct awaitable
-        {
-            awaitable(void* handle, Windows::Foundation::TimeSpan timeout) noexcept :
-                m_timeout(timeout),
-                m_handle(handle)
-            {}
-
-            bool await_ready() const noexcept
-            {
-                return WINRT_WaitForSingleObject(m_handle, 0) == 0;
-            }
-
-            void await_suspend(std::experimental::coroutine_handle<> resume)
-            {
-                m_resume = resume;
-                m_wait.attach(check_pointer(WINRT_CreateThreadpoolWait(callback, this, nullptr)));
-                int64_t relative_count = -m_timeout.count();
-                int64_t* file_time = relative_count != 0 ? &relative_count : nullptr;
-                WINRT_SetThreadpoolWait(m_wait.get(), m_handle, file_time);
-            }
-
-            bool await_resume() const noexcept
-            {
-                return m_result == 0;
-            }
-
-        private:
-
-            static void WINRT_CALL callback(void*, void* context, void*, uint32_t result) noexcept
-            {
-                auto that = static_cast<awaitable*>(context);
-                that->m_result = result;
-                that->m_resume();
-            }
-
-            struct wait_traits
-            {
-                using type = impl::ptp_wait;
-
-                static void close(type value) noexcept
-                {
-                    WINRT_CloseThreadpoolWait(value);
-                }
-
-                static constexpr type invalid() noexcept
-                {
-                    return nullptr;
-                }
-            };
-
-            handle_type<wait_traits> m_wait;
-            Windows::Foundation::TimeSpan m_timeout;
-            void* m_handle;
-            uint32_t m_result{};
-            std::experimental::coroutine_handle<> m_resume{ nullptr };
-        };
-
-        return awaitable{ handle, timeout };
-    }
-
-#ifdef _RESUMABLE_FUNCTIONS_SUPPORTED
-    inline auto operator co_await(Windows::Foundation::TimeSpan duration)
-    {
-        return resume_after(duration);
-    }
-#endif
-
     struct get_progress_token_t {};
 
     inline get_progress_token_t get_progress_token() noexcept
