+++ src/tool/cpp/cppwinrt/strings/base_coroutine_action_with_progress.h
