+++ src/tool/cpp/cppwinrt/strings/base_coroutine_operation.h
