+++ src/tool/cpp/cppwinrt/strings/base_coroutine_operation_with_progress.h
